ÐИсправления замечаний
