+++ src/components/Footer/index.js
@@ -1,10 +1,10 @@
-import ('./Footer.css');
+import s from './style.module.css';
 
 const Footer = () => {
     return(
         <>
             <footer>
-                <div className="wrapper">
+                <div className={s.wrapper}>
                     <h3>THANKS FOR VISITING</h3>
                     <p>© 2021 #ReactMarathon.</p>
                 </div>
