+++ src/components/Header/index.js
@@ -1,11 +1,11 @@
-import './Header.css';
+import s from './style.module.css';
 
 const Header = ({title, descr}) => {
     return (
         <>
-            <header className="root">
-                <div className="forest"></div>
-                <div className="container">
+            <header className={s.root}>
+                <div className={s.forest}></div>
+                <div className={s.container}>
                     <h1>{title}</h1>
                     <p>{descr}</p>
                 </div>
