+++ src/components/Layout/index.js
@@ -1,0 +1,27 @@
+import s from './style.module.css';
+
+const Layout = ({title, descr, colorBg, urlBg}) => {
+    const layoutStyle = {
+        backgroundColor: colorBg || '',
+        backgroundImage: urlBg ? `url('${urlBg}')` : ''
+    };
+    return (
+        <>
+            <section className={s.root} style={layoutStyle}>
+                <div className={s.wrapper}>
+                    <article>
+                        <div className={s.title}>
+                            <h3>{title}</h3>
+                            <span className={s.separator} />
+                        </div>
+                        <div className={`${s.desc} ${s.full}`}>
+                            <p>{descr}</p>
+                        </div>
+                    </article>
+                </div>
+            </section>
+        </>
+    )
+}
+
+export default Layout
