+++ src/components/Layout/style.module.css
@@ -7,6 +7,7 @@
     clip-path: polygon(0 0, 20% 5%, 100% 0, 100% 100%, 80% 95%, 0 100%);
     z-index: 5;
     background-attachment: fixed;
+    color: white;
 }
 
 .root .wrapper {
